+++ syserol/regression.py
@@ -2,7 +2,6 @@
 from functools import reduce
 import numpy as np
 import pandas as pd
-
 from sklearn.linear_model import ElasticNetCV, ElasticNet, LogisticRegressionCV
 from sklearn.model_selection import cross_val_predict
 from sklearn.metrics import r2_score, confusion_matrix, accuracy_score
@@ -15,6 +14,7 @@ from .dataImport import (
     importIGG,
     load_file,
     importAlterDF,
+    getAxes,
 )
 from .tensor import perform_CMTF
 
@@ -71,7 +71,18 @@ def function_elastic_net(function="ADCC"):
 def two_way_classifications():
     """ Predict classifications of subjects by progression (EC/VC vs TP/UP) or by viremia (EC/TP vs VC/UP) - Alter methods"""
     # Import Luminex, Luminex-IGG, Subject group pairs, and Glycan into DF
-    df_merged = importAlterDF()
+    df = importLuminex()
+    lum = df.pivot(index="subject", columns="variable", values="value")
+    subj = load_file("meta-subjects")
+    igg = importIGG()
+
+    igg = igg.pivot(index="subject", columns="variable", values="value")
+    data_frames = [lum, subj, igg]
+    df_merged = reduce(
+        lambda left, right: pd.merge(left, right, on=["subject"], how="inner"),
+        data_frames,
+    )
+    df_merged = df_merged.dropna()
 
     # Subset, Z score
     df_class = df_merged[["class.cp", "class.nv"]]
@@ -107,3 +118,67 @@ def two_way_classifications():
     print(f"Confusion Matrix Viremic vs. Nonviremic: {confusionVvN} \n")
 
     return accuracyCvP, accuracyVvN, confusionCvP, confusionVvN
+
+
+def noCMTF_function_prediction(components=6, function="ADCC"):
+    """ Predict functions using our decomposition and regression methods"""
+    cube, glyCube = createCube()
+    tensorFac, matrixFac, _ = perform_CMTF(cube, glyCube, components)
+
+    func, _ = importFunction()
+    df = pd.DataFrame(tensorFac[1][0])  # subjects x components matrix
+    df = df.join(func, how="inner")
+    df = df.dropna()
+    df_func = df[["ADCD", "ADCC", "ADNP", "CD107a", "IFNy", "MIP1b"]]
+    df_variables = df.drop(
+        ["subject", "ADCD", "ADCC", "ADNP", "CD107a", "IFNy", "MIP1b"], axis=1
+    )
+
+    X = df_variables
+    Y = df_func[function]
+    regr = ElasticNetCV(normalize=True, max_iter=10000)
+    model = regr.fit(X, Y)
+    Y_pred = cross_val_predict(
+        ElasticNet(alpha=regr.alpha_, normalize=True, max_iter=10000), X, Y, cv=10
+    )
+
+    print(f"Components: {components}, Accuracy: {np.sqrt(r2_score(Y, Y_pred))}")
+
+    return Y, Y_pred, np.sqrt(r2_score(Y, Y_pred))
+
+
+def ourSubjects_function_prediction(components=6, function="ADCC"):
+    """ Predict functions for subjects specifically left out of Alter using regression methods"""
+    # Re-Create Alter DataFrame with leftout subjects
+    df_merged = importAlterDF()
+
+    fullsubj = np.array(df_merged["subject"])  # Subjects only included in Alter
+    leftout = []
+    subjects, _, _ = getAxes()
+    for index, i in enumerate(subjects):
+        if i not in fullsubj:
+            leftout.append((index, i))  # Subjects left out of Alter
+    indices = [i[0] for i in leftout]
+
+    cube, glyCube = createCube()
+    tensorFac, matrixFac, _ = perform_CMTF(cube, glyCube, components)
+
+    func, _ = importFunction()
+    df = pd.DataFrame(tensorFac[1][0])  # subjects x components matrix
+    df = df.join(func, how="inner")
+    df = df.iloc[indices]
+    df = df.dropna()
+    df_func = df[["ADCD", "ADCC", "ADNP", "CD107a", "IFNy", "MIP1b"]]
+    df_variables = df.drop(
+        ["subject", "ADCD", "ADCC", "ADNP", "CD107a", "IFNy", "MIP1b"], axis=1
+    )
+
+    X = df_variables
+    Y = df_func[function]
+    regr = ElasticNetCV(normalize=True, max_iter=10000)
+    model = regr.fit(X, Y)
+    Y_pred = cross_val_predict(
+        ElasticNet(alpha=regr.alpha_, normalize=True, max_iter=10000), X, Y, cv=10
+    )
+
+    return Y, Y_pred
